+++ gsrest/service/bulk_service.py
@@ -1,5 +1,5 @@
 import importlib
-from csv import DictWriter
+from csv import DictWriter, Error as CSVError
 from openapi_server.models.values import Values
 from openapi_server.models.entity import Entity
 from openapi_server.models.address_tag import AddressTag
@@ -8,6 +8,13 @@ import json
 from aiohttp import web
 import traceback
 import inspect
+import contextlib
+from functools import reduce
+
+
+@contextlib.asynccontextmanager
+async def asyncnullcontext():
+    yield None
 
 
 def bulk_json(*args, **kwargs):
@@ -38,10 +45,10 @@ async def bulk(request, currency, operation, body, num_pages, form='csv'):
         raise web.HTTPBadRequest(text=text)
 
     if form == 'csv':
-        gen = to_csv(the_stack)
+        gen = to_csv(the_stack, request.app.logger)
         mimetype = 'text/csv'
     else:
-        gen = to_json(the_stack)
+        gen = to_json(the_stack, request.app.logger)
         mimetype = 'application/json'
 
     response = web.StreamResponse(status=200,
@@ -99,21 +106,25 @@ def flatten(item, name="", flat_dict=None, format=None):
             name = name[:-1]
             item = [i if isinstance(i, str) else str(i) for i in item if i]
             flat_dict[name] = ','.join(item)
-            flat_dict[f'{name}_count'] = len(item)
+            if not name == "actors":
+                flat_dict[f'{name}_count'] = len(item)
         else:
             flat_dict[name[:-1]] = \
                 [flatten(sub_item, format=format) for sub_item in item]
     else:
+        # if item is not None:
         flat_dict[name[:-1]] = item
     return flat_dict
 
 
-async def wrap(request, operation, currency, params, keys, num_pages, format):
+async def wrap(request, operation, currency, params, keys, num_pages, format,
+               max_concurrency_sem_context):
     params = dict(params)
     for (k, v) in keys.items():
         params[k] = v
     try:
-        result = await operation(request, currency, **params)
+        async with max_concurrency_sem_context:
+            result = await operation(request, currency, **params)
     except RuntimeError:
         result = {error_field: 'not found'}
     except ValueError as e:
@@ -157,7 +168,7 @@ async def wrap(request, operation, currency, params, keys, num_pages, format):
     if num_pages > 0 and page_state:
         params['page'] = page_state
         more = await wrap(request, operation, currency, params, keys,
-                          num_pages, format)
+                          num_pages, format, max_concurrency_sem_context)
         for row in more:
             flat.append(row)
     return flat
@@ -168,6 +179,7 @@ def stack(request, currency, operation, body, num_pages, format):
         for api in apis:
             mod = importlib.import_module(f'gsrest.service.{api}_service')
             if hasattr(mod, operation):
+                operation_name = operation
                 operation = getattr(mod, operation)
                 break
     except ModuleNotFoundError:
@@ -176,6 +188,9 @@ def stack(request, currency, operation, body, num_pages, format):
         raise RuntimeError(f'{api}.{operation}'
                            ' not found')
     aws = []
+
+    max_concurrency_bulk_operation = request.app['config']['database'].get(
+        f"max_concurrency_bulk_{operation_name}", 10)
 
     params = {}
     keys = {}
@@ -199,77 +214,120 @@ def stack(request, currency, operation, body, num_pages, format):
         raise TypeError('Keys need to be passed as list')
     inspect.getcallargs(operation, **check)
 
+    if operation_name in ["list_entity_txs", "list_entity_addresses"]:
+        context = asyncio.Semaphore(max_concurrency_bulk_operation)
+    else:
+        # async null context, does nothing
+        # https://stackoverflow.com/questions/61479059
+        # /it-there-any-default-asynchronious-null-context-manager-in-python3-7
+        context = contextlib.AsyncExitStack()
+
     for i in range(0, ln):
         the_keys = {}
         for (k, v) in keys.items():
             the_keys[k] = v[i]
         aw = wrap(request, operation, currency, params, the_keys, num_pages,
-                  format)
+                  format, context)
 
         aws.append(aw)
 
     return asyncio.as_completed(aws)
 
 
-async def to_csv(stack):
+async def to_csv(stack, logger):
     wr = writer()
-    csv = None
 
-    stash = []
-    has_data = False
+    def is_count_column(row, key):
+        postfix = "_count"
+        return key.endswith(postfix) and key[:-len(postfix)] in row and type(
+            row.get(key, None)) == int
 
+    def write_csv_row(csvwriter, buffer_writer, row, header_columns):
+        try:
+            # Filter rows not in header rows
+            # (otherwise we get an error, unexpected keys)
+            # This error is easy to go unnoticed.
+            # One situation where such an
+            # error is triggered are optional lists
+            # (lets call it A) fields,
+            # since the first row the function flatten
+            # adds an A_count field for
+            # every list if the list field is optional
+            # it can be none,
+            # if the first row has a none value for that
+            # optional field the type
+            # is not list thus no A_count field is added.
+            # Because the first row is
+            # used to create the headerline for the csv
+            # the A_count is not part of the header
+            # but if another row has a value set for A
+            # (e.g. type(A)==list) a count field
+            # is added. This causes problems since we write the
+            # data into a stream
+            # thus the header is already out. Also the DictWriter
+            # raises a value error
+            # on unexpected fields (only count_columns)
+            out_row = {
+                k: v
+                for k, v in row.items()
+                if (k in header_columns or not is_count_column(row, k))
+            }
+
+            csvwriter.writerow(out_row)
+        except (ValueError, CSVError) as e:
+            logger.error(f"Error writing bulk row {row}: ({type(e)}) {e}")
+            request_fields = {
+                k: v
+                for k, v in row.items() if k.startswith(request_field_prefix)
+            }
+            error_and_request_fields = {
+                **{
+                    error_field: "internal error - can't produce csv"
+                },
+                **request_fields
+            }
+            csvwriter.writerow(error_and_request_fields)
+        return buffer_writer.get()
+
+    NR_REGULAR_ROWS_USED_TO_INFER_HEADER = 100
+    rows_to_infer_header = []
+    regular_rows = 0
+    ops_rest = []
     for op in stack:
+        if regular_rows < NR_REGULAR_ROWS_USED_TO_INFER_HEADER:
+            rows = await op
+            rows_to_infer_header.extend(rows)
+            regular_rows += sum(
+                1 for r in rows
+                if info_field not in r and error_field not in r)
+        else:
+            ops_rest.append(op)
+
+    # Infer header
+    headerfields = sorted(
+        list(
+            reduce(set.union, [set(r.keys()) for r in rows_to_infer_header],
+                   set()).union(set([error_field, info_field]))))
+
+    csv = DictWriter(wr, headerfields, restval='', extrasaction='ignore')
+
+    # write header
+    csv.writeheader()
+    head = wr.get()
+    yield head
+
+    # write header infer rows
+    for row in rows_to_infer_header:
+        yield write_csv_row(csv, wr, row, headerfields)
+
+    # write the rest
+    for op in ops_rest:
         rows = await op
-
         for row in rows:
-
-            if error_field in row and not csv:
-                stash.append(row)
-                continue
-
-            if info_field in row and not csv:
-                stash.append(row)
-                continue
-
-            head = ""
-            has_data = True
-            if not csv:
-                fieldnames = list(row.keys())
-                fieldnames.append(info_field)
-                fieldnames.append(error_field)
-                fieldnames = sorted(fieldnames)
-                csv = DictWriter(wr, fieldnames)
-                csv.writeheader()
-                head = wr.get()
-                yield head
-
-            for stashed_row in stash:
-                csv.writerow(stashed_row)
-                yield wr.get()
-
-            stash = []
-
-            csv.writerow(row)
-            yield wr.get()
-
-    if not has_data:
-        for row in stash:
-            if not csv:
-                fieldnames = list(row.keys())
-                if info_field not in fieldnames:
-                    fieldnames.append(info_field)
-                if error_field not in fieldnames:
-                    fieldnames.append(error_field)
-                fieldnames = sorted(fieldnames)
-                csv = DictWriter(wr, fieldnames)
-                csv.writeheader()
-                yield wr.get()
-
-            csv.writerow(row)
-            yield wr.get()
+            yield write_csv_row(csv, wr, row, headerfields)
 
 
-async def to_json(stack):
+async def to_json(stack, logger):
     started = False
     yield "["
     for op in stack: